Update Lesson 2 README and starter code
This commit updates the README for Lesson 2 to provide more detailed instructions on navigating the "L" shaped course and introduces the concept of encoders.

It also updates the `Lesson2_encoder_based_starter.java` file to include commented-out code for moving the robot forward to the flag, guiding students to complete the implementation.

Additionally, a minor typo in the Lesson 1 README link to the starter code has been corrected.